+++ src/domain/search/index.ts
@@ -10,7 +10,6 @@ export type { ModelConfig } from './models.js';
 export { DEFAULT_MODEL, disposeModel, EMBEDDING_STRATEGIES, embed, MODELS } from './models.js';
 export type { RemoteEmbeddingOptions } from './providers/remote.js';
 export { embedRemote, resolveRemoteEmbeddingOptions } from './providers/remote.js';
-export { search } from './search/cli-formatter.js';
 export { hybridSearchData } from './search/hybrid.js';
 export { ftsSearchData } from './search/keyword.js';
 export { multiSearchData, searchData } from './search/semantic.js';
